+++ core/modules/modmassaction.class.php
@@ -110,7 +110,7 @@ class modmassaction extends DolibarrModules
 		$this->requiredby = array();	// List of modules id to disable if this one is disabled
 		$this->conflictwith = array();	// List of modules id this module is in conflict with
 		$this->phpmin = array(7,0);					// Minimum version of PHP required by module
-		$this->need_dolibarr_version = array(15,0);	// Minimum version of Dolibarr required by module
+		$this->need_dolibarr_version = array(16,0);	// Minimum version of Dolibarr required by module
 		$this->langfiles = array("massaction@massaction");
 
 		// Constants
@@ -147,7 +147,7 @@ class modmassaction extends DolibarrModules
         $this->tabs = array();
 
         // Dictionaries
-	    if (! isset($conf->massaction->enabled))
+	    if (! isModEnabled('massaction'))
         {
         	$conf->massaction=new stdClass();
         	$conf->massaction->enabled=0;