+++ ChangeLog.md
@@ -7,6 +7,7 @@
 - NEW : COMPAT V20 - *26/07/2024* - 1.6.0
 
 ## Release 1.5
+- FIX : DA026002 Select only checkbox with checkforselect class to reset checked - *05/02/2024* - 1.5.2
 - FIX : ICON siccors  in > 18.0  - *17/06/2024* - 1.5.1
   - remove question  in massaction confirm
   - set main handler in hook section