+++ ChangeLog.md
@@ -3,11 +3,13 @@
 ## Unreleased
 
 ## Release 1.6
-- FIX : DA026579 => fix undeclared variable toShow  - *23/05/2025* - 1.6.2
+- FIX : DA026579 => fix undeclared variable toShow  - *23/05/2025* - 1.6.3
+- FIX : Using $this in non object context - *15/04/2025* - 1.6.2
 - FIX : COMPAT V21 - *06/12/2024* - 1.6.1
 - NEW : COMPAT V20 - *26/07/2024* - 1.6.0
 
 ## Release 1.5
+- FIX : DA026002 Select only checkbox with checkforselect class to reset checked - *05/02/2024* - 1.5.2
 - FIX : ICON siccors  in > 18.0  - *17/06/2024* - 1.5.1
   - remove question  in massaction confirm
   - set main handler in hook section